Improve code readability, simplicity, and add search feature
Added Packages:
   - Brought in `use-debounce` to make our search functionality smoother.
   - Introduced `heroicons` for a consistent look across the site.
   - Also, added `clsx` for cleaner handling of classes.

Worked on restructuring the code to make it cleaner and easier to understand. Created a new component, `Search.jsx`, that takes care of database queries using React hooks and search parameters. Overhauled the logic in `gallery/page.js` to work nicer with pagination and new search bar.
Added a loading component from Next.js to show users when data is in the process of being fetched.